split projects and pages from same file and delete user model&controler and set database to quicapp
