+++ .env
@@ -1,2 +1,1 @@
-MONGODB_URI=mongodb://localhost:27017/test
-SECRET_TOKEN=catswillruletheworld
+MONGODB_URI=mongodb://localhost:27017/quicapp
