+++ client/app/modules/projects/models/page.interface.ts
@@ -1,0 +1,7 @@
+export interface iPage{
+     _id: String,
+    name: String,
+    layout: String,
+    creationTime: Date,
+    parent: String
+}
