+++ server/models/project.ts
@@ -1,5 +1,6 @@
 import * as mongoose from 'mongoose';
 import { Project, Page } from "../../common/pojo/project";
+import  PageModel   from './page';
 
 const projectSchema = new mongoose.Schema({
     name: String,
@@ -13,7 +14,7 @@ projectSchema.post('remove', function (doc) {
     console.log('Deleting Project', doc.id);
     PageModel.find({ parent: doc.id }, (err, docs) => {
         if (err) { return console.error(err); }
-        for (let i = 0; i < docs.length; i++) {
+        for(const [i, doc] of docs.entries()){
             console.log('Deleting Page', docs[i].id);
             docs[i].remove();
         }
@@ -22,13 +23,4 @@ projectSchema.post('remove', function (doc) {
 
 const ProjectModel = mongoose.model('Project', projectSchema);
 
-const pageSchema = new mongoose.Schema({
-    name: String,
-    layout: String,
-    creationTime: Date,
-    parent: String
-});
-
-const PageModel = mongoose.model('Page', pageSchema);
-
-export default { ProjectModel, PageModel };
+export default  { ProjectModel };
